Extract ITokenConfiguration and move ToTokenTree to extensions
Introduced ITokenConfiguration interface and moved ToTokenTree logic to TokenConfigurationExtensions.
Updated TokenReaderStateMachine to use ITokenConfiguration.
Refactored Tokenizer.EnumerateTokensAsync to accept TokenReaderStateMachine directly.
Removed indexer from TokenConfiguration.
