+++ Tokensharp/ITokenConfiguration.cs
@@ -1,0 +1,6 @@
+﻿namespace Tokensharp;
+
+public interface ITokenConfiguration<TTokenType> : IEnumerable<LexemeToTokenType<TTokenType>>
+    where TTokenType : TokenType<TTokenType>, ITokenType<TTokenType>
+{
+}
